+++ connect/amazon/sqs/fake/src/main/kotlin/org/http4k/connect/amazon/sqs/FakeSQS.kt
@@ -14,25 +14,32 @@ import org.http4k.connect.storage.Storage
 import org.http4k.core.Method.POST
 import org.http4k.routing.bind
 import org.http4k.routing.routes
+import java.time.Clock
+import java.util.concurrent.atomic.AtomicLong
 
 class FakeSQS(
     private val queues: Storage<List<SQSMessage>> = Storage.InMemory(),
     awsAccount: AwsAccount = AwsAccount.of("1234567890"),
-    private val region: Region = Region.of("ldn-north-1")
+    private val region: Region = Region.of("ldn-north-1"),
+    private val deduplication: Storage<DeduplicationRecord> = Storage.InMemory(),
+    private val queueConfig: Storage<QueueConfig> = Storage.InMemory(),
+    private val clock: Clock = Clock.systemUTC()
 ) : ChaoticHttpHandler() {
 
     private val api = AwsRestJsonFake(SqsMoshi, AwsService.of("sqs"), region, awsAccount)
 
+    private val sequenceNumbers = AtomicLong()
+
     override val app = routes(
         "/" bind POST to routes(
             api.deleteMessage(queues),
-            api.deleteQueue(queues),
+            api.deleteQueue(queues, deduplication, queueConfig),
             api.deleteMessageBatch(queues),
             api.receiveMessage(queues),
-            api.createQueue(queues, awsAccount),
+            api.createQueue(queues, queueConfig, awsAccount),
             api.getQueueAttributes(queues),
-            api.sendMessage(queues),
-            api.sendMessageBatch(queues),
+            api.sendMessage(queues, deduplication, queueConfig, clock, sequenceNumbers),
+            api.sendMessageBatch(queues, deduplication, queueConfig, clock, sequenceNumbers),
             api.listQueues(region, awsAccount, queues)
         )
     )